:recycle: (Header) create emptyHeader and use it in subscription
